(to squash) - move existing config definition into separate file
