+++ src/core/BUILD.gn
@@ -590,6 +590,8 @@ openthread_core_files = [
   "thread/network_data_service.cpp",
   "thread/network_data_service.hpp",
   "thread/network_data_tlvs.hpp",
+  "thread/network_data_types.cpp",
+  "thread/network_data_types.hpp",
   "thread/network_diagnostic.cpp",
   "thread/network_diagnostic.hpp",
   "thread/network_diagnostic_tlvs.hpp",
