+++ src/core/CMakeLists.txt
@@ -193,6 +193,7 @@ set(COMMON_SOURCES
     thread/network_data_local.cpp
     thread/network_data_notifier.cpp
     thread/network_data_service.cpp
+    thread/network_data_types.cpp
     thread/network_diagnostic.cpp
     thread/panid_query_server.cpp
     thread/radio_selector.cpp
