+++ src/core/thread/network_data.cpp
@@ -47,73 +47,6 @@
 namespace ot {
 namespace NetworkData {
 
-void OnMeshPrefixConfig::SetFrom(const PrefixTlv &        aPrefixTlv,
-                                 const BorderRouterTlv &  aBorderRouterTlv,
-                                 const BorderRouterEntry &aBorderRouterEntry)
-{
-    Clear();
-
-    aPrefixTlv.CopyPrefixTo(GetPrefix());
-    mPreference   = aBorderRouterEntry.GetPreference();
-    mPreferred    = aBorderRouterEntry.IsPreferred();
-    mSlaac        = aBorderRouterEntry.IsSlaac();
-    mDhcp         = aBorderRouterEntry.IsDhcp();
-    mConfigure    = aBorderRouterEntry.IsConfigure();
-    mDefaultRoute = aBorderRouterEntry.IsDefaultRoute();
-    mOnMesh       = aBorderRouterEntry.IsOnMesh();
-    mStable       = aBorderRouterTlv.IsStable();
-    mRloc16       = aBorderRouterEntry.GetRloc();
-    mNdDns        = aBorderRouterEntry.IsNdDns();
-    mDp           = aBorderRouterEntry.IsDp();
-}
-
-void ExternalRouteConfig::SetFrom(Instance &           aInstance,
-                                  const PrefixTlv &    aPrefixTlv,
-                                  const HasRouteTlv &  aHasRouteTlv,
-                                  const HasRouteEntry &aHasRouteEntry)
-{
-    Clear();
-
-    aPrefixTlv.CopyPrefixTo(GetPrefix());
-    mPreference          = aHasRouteEntry.GetPreference();
-    mNat64               = aHasRouteEntry.IsNat64();
-    mStable              = aHasRouteTlv.IsStable();
-    mRloc16              = aHasRouteEntry.GetRloc();
-    mNextHopIsThisDevice = (aHasRouteEntry.GetRloc() == aInstance.Get<Mle::MleRouter>().GetRloc16());
-}
-
-bool ServiceConfig::ServerConfig::operator==(const ServerConfig &aOther) const
-{
-    return (mStable == aOther.mStable) && (mServerDataLength == aOther.mServerDataLength) &&
-           (memcmp(mServerData, aOther.mServerData, mServerDataLength) == 0);
-}
-
-void ServiceConfig::ServerConfig::SetFrom(const ServerTlv &aServerTlv)
-{
-    mStable           = aServerTlv.IsStable();
-    mRloc16           = aServerTlv.GetServer16();
-    mServerDataLength = aServerTlv.GetServerDataLength();
-    memcpy(&mServerData, aServerTlv.GetServerData(), mServerDataLength);
-}
-
-bool ServiceConfig::operator==(const ServiceConfig &aOther) const
-{
-    return (mEnterpriseNumber == aOther.mEnterpriseNumber) && (mServiceDataLength == aOther.mServiceDataLength) &&
-           (memcmp(mServiceData, aOther.mServiceData, mServiceDataLength) == 0) &&
-           (GetServerConfig() == aOther.GetServerConfig());
-}
-
-void ServiceConfig::SetFrom(const ServiceTlv &aServiceTlv, const ServerTlv &aServerTlv)
-{
-    Clear();
-
-    mServiceId         = aServiceTlv.GetServiceId();
-    mEnterpriseNumber  = aServiceTlv.GetEnterpriseNumber();
-    mServiceDataLength = aServiceTlv.GetServiceDataLength();
-    memcpy(&mServiceData, aServiceTlv.GetServiceData(), mServiceDataLength);
-    GetServerConfig().SetFrom(aServerTlv);
-}
-
 NetworkData::NetworkData(Instance &aInstance, Type aType)
     : InstanceLocator(aInstance)
     , mType(aType)
