+++ src/core/thread/network_data.hpp
@@ -48,6 +48,7 @@
 #include "thread/lowpan.hpp"
 #include "thread/mle_router.hpp"
 #include "thread/network_data_tlvs.hpp"
+#include "thread/network_data_types.hpp"
 
 namespace ot {
 
@@ -100,145 +101,6 @@ enum
  *
  */
 typedef otNetworkDataIterator Iterator;
-
-/**
- * This class represents an On Mesh Prefix (Border Router) configuration.
- *
- */
-class OnMeshPrefixConfig : public otBorderRouterConfig,
-                           public Clearable<OnMeshPrefixConfig>,
-                           public Equatable<OnMeshPrefixConfig>
-{
-    friend class NetworkData;
-
-public:
-    /**
-     * This method gets the prefix.
-     *
-     * @return The prefix.
-     *
-     */
-    const Ip6::Prefix &GetPrefix(void) const { return static_cast<const Ip6::Prefix &>(mPrefix); }
-
-    /**
-     * This method gets the prefix.
-     *
-     * @return The prefix.
-     *
-     */
-    Ip6::Prefix &GetPrefix(void) { return static_cast<Ip6::Prefix &>(mPrefix); }
-
-private:
-    void SetFrom(const PrefixTlv &        aPrefixTlv,
-                 const BorderRouterTlv &  aBorderRouterTlv,
-                 const BorderRouterEntry &aBorderRouterEntry);
-};
-
-/**
- * This class represents an External Route configuration.
- *
- */
-class ExternalRouteConfig : public otExternalRouteConfig,
-                            public Clearable<ExternalRouteConfig>,
-                            public Equatable<ExternalRouteConfig>
-{
-    friend class NetworkData;
-
-public:
-    /**
-     * This method gets the prefix.
-     *
-     * @return The prefix.
-     *
-     */
-    const Ip6::Prefix &GetPrefix(void) const { return static_cast<const Ip6::Prefix &>(mPrefix); }
-
-    /**
-     * This method gets the prefix.
-     *
-     * @return The prefix.
-     *
-     */
-    Ip6::Prefix &GetPrefix(void) { return static_cast<Ip6::Prefix &>(mPrefix); }
-
-    /**
-     * This method sets the prefix.
-     *
-     * @param[in]  aPrefix  The prefix to set to.
-     *
-     */
-    void SetPrefix(const Ip6::Prefix &aPrefix) { mPrefix = aPrefix; }
-
-private:
-    void SetFrom(Instance &           aInstance,
-                 const PrefixTlv &    aPrefixTlv,
-                 const HasRouteTlv &  aHasRouteTlv,
-                 const HasRouteEntry &aHasRouteEntry);
-};
-
-/**
- * This type represents a Service configuration.
- *
- */
-class ServiceConfig : public otServiceConfig, public Clearable<ServiceConfig>, public Unequatable<ServiceConfig>
-{
-    friend class NetworkData;
-
-public:
-    /**
-     * This class represents a Server configuration.
-     *
-     */
-    class ServerConfig : public otServerConfig, public Unequatable<ServerConfig>
-    {
-        friend class ServiceConfig;
-
-    public:
-        /**
-         * This method overloads operator `==` to evaluate whether or not two `ServerConfig` instances are equal.
-         *
-         * @param[in]  aOther  The other `ServerConfig` instance to compare with.
-         *
-         * @retval TRUE   If the two `ServerConfig` instances are equal.
-         * @retval FALSE  If the two `ServerConfig` instances are not equal.
-         *
-         */
-        bool operator==(const ServerConfig &aOther) const;
-
-    private:
-        void SetFrom(const ServerTlv &aServerTlv);
-    };
-
-    /**
-     * This method gets the Server configuration.
-     *
-     * @returns The Server configuration.
-     *
-     */
-    const ServerConfig &GetServerConfig(void) const { return static_cast<const ServerConfig &>(mServerConfig); }
-
-    /**
-     * This method gets the Server configuration.
-     *
-     * @returns The Server configuration.
-     *
-     */
-    ServerConfig &GetServerConfig(void) { return static_cast<ServerConfig &>(mServerConfig); }
-
-    /**
-     * This method overloads operator `==` to evaluate whether or not two `ServiceConfig` instances are equal.
-     *
-     * @param[in]  aOther  The other `ServiceConfig` instance to compare with.
-     *
-     * @retval TRUE   If the two `ServiceConfig` instances are equal.
-     * @retval FALSE  If the two `ServiceConfig` instances are not equal.
-     *
-     */
-    bool operator==(const ServiceConfig &aOther) const;
-
-private:
-    void SetFrom(const ServiceTlv &aServiceTlv, const ServerTlv &aServerTlv);
-};
 
 /**
  * This class implements Network Data processing.
